Cleaned up auth setup (removed unnecessary code and files)
